refactor: move types and classes to their own files
Seperate each declaration and export them from their own file for more
