+++ src/types/ButtonStyle.ts
@@ -1,0 +1,4 @@
+/**
+ * The style of the paginator buttons.
+ */
+export type ButtonStyle = "PRIMARY" | "SECONDARY" | "DANGER" | "SUCCESS";
