+++ src/types/LabelOptions.ts
@@ -1,0 +1,24 @@
+//TODO: add docs to describe interface
+
+export interface LabelOptions {
+  /**
+   * The label to use for the first button.
+   * @default ""
+   */
+  firstLabel: string;
+  /**
+   * The label to use for the previous button.
+   * @default ""
+   */
+  prevLabel: string;
+  /**
+   * The label to use for the next button.
+   * @default ""
+   */
+  nextLabel: string;
+  /**
+   * The label to use for the last button.
+   * @default ""
+   */
+  lastLabel: string;
+}
