+++ src/types/index.ts
@@ -1,0 +1,6 @@
+export * from "./ButtonOptions";
+export * from "./ButtonsOptions";
+export * from "./ButtonStyle";
+export * from "./EmojiOptions";
+export * from "./LabelOptions";
+export * from "./Options";
